requirements.txt
streamlit==1.35.0
PyYAML==6.0.1
Jinja2==3.1.4